backend/pyproject.toml
[project]
name = "subscriptions-backend"
version = "0.1.0"
description = "Subscriptions tracking backend service"
authors = [
    {name = "Subscriptions Team"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "pydantic-settings (>=2.6.1,<3.0.0)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "psycopg[binary] (>=3.1.18,<4.0.0)",
    "alembic (>=1.13.3,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
